Split TicketState into its own file
