+++ src/components/Ticket.css
@@ -17,17 +17,3 @@
 	margin-left: 2%;
 	flex-shrink: 0;
 }
-
-.Ticket-state {
-	border-radius: 2px;
-	background: #82878c;
-	color: #fff;
-	padding: 0.15em 0.4em;
-	margin-right: 0.5em;
-}
-.Ticket-state.state-open {
-	background: #46b450;
-}
-.Ticket-state.state-closed {
-	background: #dc3232;
-}
